Fix error in delete idea.
In the current implementation of CmQuiz::Review::DeleteIdea; after an idea is deleted, a call is made with the get ideas api that is expected to not return any idea.
This is however incorrect because get idea pulls top ten ideas by average score(or less than 10 if ideas are less than 10).
Since previously created ideas are not deleted, this spec will always fail because get_idea will return ideas created before the just deleted idea, hence, the size of its response can not be zero.

This commit fixes that.