Fix bug in CmQuiz::Review::GetIdeas
The current implementation of this class will only make the test pass a few times. Say one or two.
Once the number of ideas created is more than 10, it begins to fail.
This behaviour is also caused because we do not delete previously created records from this test api.

The fixes the issue and regardless of the number of times the test is run, this spec should not fail.