Ensure a deleted idea does not exist in GetIdeas response anymore
In the old way, if some ideas are already there in the database,
checking correctness of deleting an idea by checking equality of
the size of returned ideas from GetIdeas endpoint to be ZERO will
always fail, whether the idea is properly deleted or not.

This commit asserts that the deleted idea does not exist in the
resulting array of ideas anymore.